requirements.txt
mysql==0.0.3
mysql-connector-python==8.3.0
mysqlclient==2.2.
asyncio==3.4.3
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
colorama==0.4.4
dnspython==2.1.0
Flask==2.0.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
python-dotenv==0.19.2
requests==2.26.0
urllib3==1.26.7
watchdog==2.1.6
Werkzeug==2.0.2